LayoutTests/imported/w3c/web-platform-tests/wasm/serialization/module/serialization-via-notifications-api.any-expected.txt


PASS WebAssembly.Module cloning via the Notifications API's data member: basic case
PASS WebAssembly.Module cloning via the Notifications API's data member: is interleaved correctly
